chore: rename spicey-la-vicey lambda
